requirements.txt
google-auth>=2.38.0
google-auth-oauthlib>=1.2.1
browser-cookie3>=0.20.1
yt-dlp>=2025.10.14
curl-cffi>=0.13.0
